+++ src/CCCircleWave.cpp
@@ -1,4 +1,5 @@
 #include <Geode/modify/CCCircleWave.hpp>
+#include "Manager.hpp"
 #include "Utils.hpp"
 
 #define PREFERRED_HOOK_PRIO (-3999)
@@ -10,15 +11,14 @@ class $modify(MyCCCircleWave, CCCircleWave) {
 	void setPosition(cocos2d::CCPoint const& p0) {
 		CCCircleWave::setPosition(p0);
 
-		if (!Utils::modEnabled() || !Utils::getBool("hideLevelCompleteVFX") || Utils::isEclipse()) return;
+		const Manager* manager = Manager::getSharedInstance();
+		if (!manager->modEnabled || !manager->hideLevelCompleteVFX || Utils::isEclipse()) return;
 
 		// for relevant comments on the source code below this line, please see:
 		// https://raw.githubusercontent.com/EclipseMenu/EclipseMenu/main/src/hacks/Level/HideLevelCompleteVFX.cpp
 		// reused with permission on grounds of self-authored code
 
-		PlayLayer* pl = PlayLayer::get();
-
-		if (!pl || !pl->m_levelEndAnimationStarted) return;
+		if (const PlayLayer* pl = PlayLayer::get(); !pl || !pl->m_levelEndAnimationStarted) return;
 		
 		if (!geode::cast::typeinfo_cast<CurrencyRewardLayer*>(this->getParent()))
 			this->setVisible(false);