+++ src/EndPortalObject.cpp
@@ -5,15 +5,13 @@
 using namespace geode::prelude;
 
 class $modify(MyEndPortalObject, EndPortalObject) {
-	struct Fields {
-		Manager* manager = Manager::getSharedInstance();
-	};
 	void updateEndPos(bool p0) {
 		EndPortalObject::updateEndPos(p0);
 		if (!Utils::modEnabled()) return;
 		if (Utils::getBool("hideEndPortalGradient")) this->m_gradientBar->setVisible(false);
 		if (!Utils::getBool("noEndPortalParticles")) return;
-		m_fields->manager->endPortalX = this->getPositionX();
-		m_fields->manager->endPortalY = this->getPositionY();
+		Manager* manager = Manager::getSharedInstance();
+		manager->endPortalX = this->getPositionX();
+		manager->endPortalY = this->getPositionY();
 	}
 };