+++ src/PlayLayer.cpp
@@ -14,50 +14,49 @@ class $modify(MyPlayLayer, PlayLayer) {
 		(void) self.setHookPriority("PlayLayer::addObject", PREFERRED_HOOK_PRIO);
 	}
 	struct Fields {
-		Manager* manager = Manager::getSharedInstance();
 		bool performHideTwoPlayerGuide = false;
 		bool foundPlayerOrAudioEffectsLayer = false;
 		bool foundHitboxNodeTwoPlayerGuide = false;
 		bool foundHitboxNode = false;
 	};
-	bool isIn(std::vector<int> numberStorage, int id) {
+	static bool isIn(const std::vector<int>& numberStorage, const int id) {
 		return std::ranges::find(numberStorage, id) != numberStorage.end();
 	}
 	void addObject(GameObject* object) {
 		if (!Utils::modEnabled()) return PlayLayer::addObject(object);
-		std::string noGlow = Utils::getString("noGlow");
-		if (noGlow != "Ignore") {
+		if (const std::string& noGlow = Utils::getString("noGlow"); noGlow != "Ignore") {
 			object->m_hasNoGlow = noGlow == "Force Enable"; // this has to go here because putting it in setupHasCompleted causes objects at camera's starting position to already have glow enabled
 		}
-		if (isIn(m_fields->manager->particleObjects, object->m_objectID)) return PlayLayer::addObject(object);
+		const Manager* manager = Manager::getSharedInstance();
+		if (isIn(manager->particleObjects, object->m_objectID)) return PlayLayer::addObject(object);
 		// truth be told, i could've done a switch case based on enum value. however, actual objects with noTouch property would be unaffected
-		if (Utils::getBool("noGoldCoinParticles") && object->m_objectID == 142) object->m_hasParticles = false;
-		else if (Utils::getBool("noCoinParticles") && object->m_objectID == 1329) object->m_hasNoParticles = true;
-		else if (Utils::getBool("noCustomOrbParticles") && object->m_objectID == 1594) object->m_particleString = "";
-		else if (Utils::getBool("noTeleportOrbParticles") && object->m_objectID == 3027) object->m_particleString = "";
-		else if (Utils::getBool("noDashOrbParticles") && (object->m_objectID == 1704 || object->m_objectID == 1751)) object->m_particleString = "";
-		else if (Utils::getBool("noDualPortalParticles") && (object->m_objectID == 286 || object->m_objectID == 287)) object->m_particleString = "";
-		else if (Utils::getBool("noMirrorPortalParticles") && (object->m_objectID == 45 || object->m_objectID == 46)) object->m_particleString = "";
-		else if (Utils::getBool("noSizePortalParticles") && (object->m_objectID == 99 || object->m_objectID == 101)) object->m_particleString = "";
-		else if (Utils::getBool("noJumpRingParticles") && isIn(m_fields->manager->jumpRingObjects, object->m_objectID)) object->m_particleString = "";
-		else if (Utils::getBool("noJumpPadParticles") && isIn(m_fields->manager->jumpPadObjects, object->m_objectID)) object->m_particleString = "";
-		else if (Utils::getBool("noGamemodePortalParticles") && isIn(m_fields->manager->gameModePortals, object->m_objectID)) object->m_particleString = "";
-		else if (Utils::getBool("noSpeedPortalParticles") && isIn(m_fields->manager->speedPortalObjects, object->m_objectID)) object->m_particleString = "";
-		else if (Utils::getBool("noGravityPortalParticles") && isIn(m_fields->manager->gravityPortals, object->m_objectID)) object->m_particleString = "";
-		else if (Utils::getBool("noTeleportPortalParticles") && isIn(m_fields->manager->teleportPortalObjects, object->m_objectID)) object->m_particleString = "";
-		else if (Utils::getBool("noCollectibleParticles") && isIn(m_fields->manager->collectibleObjects, object->m_objectID)) object->m_particleString = "";
-		else if (Utils::getBool("noParticles") && !isIn(m_fields->manager->everythingInOne, object->m_objectID)) object->m_particleString = "";
+		if (manager->noGoldCoinParticles && object->m_objectID == 142) object->m_hasParticles = false;
+		else if (manager->noCoinParticles && object->m_objectID == 1329) object->m_hasNoParticles = true;
+		else if (manager->noCustomOrbParticles && object->m_objectID == 1594) object->m_particleString = "";
+		else if (manager->noTeleportOrbParticles && object->m_objectID == 3027) object->m_particleString = "";
+		else if (manager->noDashOrbParticles && (object->m_objectID == 1704 || object->m_objectID == 1751)) object->m_particleString = "";
+		else if (manager->noDualPortalParticles && (object->m_objectID == 286 || object->m_objectID == 287)) object->m_particleString = "";
+		else if (manager->noMirrorPortalParticles && (object->m_objectID == 45 || object->m_objectID == 46)) object->m_particleString = "";
+		else if (manager->noSizePortalParticles && (object->m_objectID == 99 || object->m_objectID == 101)) object->m_particleString = "";
+		else if (manager->noJumpRingParticles && isIn(manager->jumpRingObjects, object->m_objectID)) object->m_particleString = "";
+		else if (manager->noJumpPadParticles && isIn(manager->jumpPadObjects, object->m_objectID)) object->m_particleString = "";
+		else if (manager->noGamemodePortalParticles && isIn(manager->gameModePortals, object->m_objectID)) object->m_particleString = "";
+		else if (manager->noSpeedPortalParticles && isIn(manager->speedPortalObjects, object->m_objectID)) object->m_particleString = "";
+		else if (manager->noGravityPortalParticles && isIn(manager->gravityPortals, object->m_objectID)) object->m_particleString = "";
+		else if (manager->noTeleportPortalParticles && isIn(manager->teleportPortalObjects, object->m_objectID)) object->m_particleString = "";
+		else if (manager->noCollectibleParticles && isIn(manager->collectibleObjects, object->m_objectID)) object->m_particleString = "";
+		else if (manager->noParticles && !isIn(manager->everythingInOne, object->m_objectID)) object->m_particleString = "";
 		PlayLayer::addObject(object);
 	}
 	void setupHasCompleted() {
 		if (!Utils::modEnabled() || !m_objects || !m_level || !m_levelSettings) return PlayLayer::setupHasCompleted();
 		if (!std::string(m_level->m_levelString).empty()) {
-			std::string fixRobotJump = Utils::getString("fixRobotJump");
-			std::string isFlipped = Utils::getString("isFlipped");
-			std::string enable22Changes = Utils::getString("enable22Changes");
-			std::string dynamicLevelHeight = Utils::getString("dynamicLevelHeight");
-			std::string fixGravityBug = Utils::getString("fixGravityBug");
-			std::string noTimePenalty = Utils::getString("noTimePenalty");
+			const std::string& fixRobotJump = Utils::getString("fixRobotJump");
+			const std::string& isFlipped = Utils::getString("isFlipped");
+			const std::string& enable22Changes = Utils::getString("enable22Changes");
+			const std::string& dynamicLevelHeight = Utils::getString("dynamicLevelHeight");
+			const std::string& fixGravityBug = Utils::getString("fixGravityBug");
+			const std::string& noTimePenalty = Utils::getString("noTimePenalty");
 			if (fixRobotJump != "Ignore") {
 				m_levelSettings->m_fixRobotJump = fixRobotJump == "Force Enable";
 			}
@@ -78,11 +77,11 @@ class $modify(MyPlayLayer, PlayLayer) {
 			}
 		}
 		PlayLayer::setupHasCompleted();
-		std::string dontEnter = Utils::getString("dontEnter");
-		std::string dontFade = Utils::getString("dontFade");
-		std::string noAudioScale = Utils::getString("noAudioScale");
-		std::string noEffects = Utils::getString("noEffects");
-		for (auto object : CCArrayExt<GameObject*>(m_objects)) {
+		const std::string& dontEnter = Utils::getString("dontEnter");
+		const std::string& dontFade = Utils::getString("dontFade");
+		const std::string& noAudioScale = Utils::getString("noAudioScale");
+		const std::string& noEffects = Utils::getString("noEffects");
+		for (GameObject* object : CCArrayExt<GameObject*>(m_objects)) {
 			if (dontEnter != "Ignore") {
 				object->m_ignoreEnter = dontEnter == "Force Enable";
 			}
@@ -121,28 +120,29 @@ class $modify(MyPlayLayer, PlayLayer) {
 	}
 	void postUpdate(float dt) {
 		PlayLayer::postUpdate(dt);
-		if (!Utils::modEnabled()) return;
-		if (Utils::getBool("noSpeedParticles")) {
-			for (auto node : CCArrayExt<CCNode*>(this->getChildren())) {
-				if (!m_fields->foundHitboxNode) m_fields->foundHitboxNode = node->getID() == "hitbox-node";
-				if (!m_fields->foundHitboxNode || node->getZOrder() != 100) continue;
-				if (auto particle = typeinfo_cast<CCParticleSystemQuad*>(node)) particle->setVisible(false);
+		const Manager* manager = Manager::getSharedInstance();
+		if (!manager->modEnabled) return;
+		const auto fields = m_fields.self();
+		if (manager->noSpeedParticles) {
+			for (CCNode* node : CCArrayExt<CCNode*>(this->getChildren())) {
+				if (!fields->foundHitboxNode) fields->foundHitboxNode = node->getID() == "hitbox-node";
+				if (!fields->foundHitboxNode || node->getZOrder() != 100) continue;
+				if (const auto particle = typeinfo_cast<CCParticleSystemQuad*>(node)) particle->setVisible(false);
 			}
 		}
-		auto batchLayer = this->getChildByIDRecursive("batch-layer");
-		if (Utils::getBool("noPlayerParticles") && batchLayer != nullptr) {
-			for (auto node : CCArrayExt<CCNode*>(batchLayer->getChildren())) {
+		if (CCNode* batchLayer = this->getChildByIDRecursive("batch-layer"); batchLayer && manager->noPlayerParticles) {
+			for (CCNode* node : CCArrayExt<CCNode*>(batchLayer->getChildren())) {
 				if (typeinfo_cast<CCSpriteBatchNode*>(node)) {
-					if (m_fields->foundPlayerOrAudioEffectsLayer) {
-						m_fields->foundPlayerOrAudioEffectsLayer = false;
+					if (fields->foundPlayerOrAudioEffectsLayer) {
+						fields->foundPlayerOrAudioEffectsLayer = false;
 						break;
 						// player particles are always between AudioEffectsLayer and a CCSpriteBatchNode, and some are ordered after the player. breaking early here for performance reasons
 					}
 					continue;
 				}
-				if (!m_fields->foundPlayerOrAudioEffectsLayer) m_fields->foundPlayerOrAudioEffectsLayer = (node == this->m_player1 || node == this->m_player2 || typeinfo_cast<AudioEffectsLayer*>(node));
-				if (!m_fields->foundPlayerOrAudioEffectsLayer) continue;
-				auto particle = typeinfo_cast<CCParticleSystemQuad*>(node);
+				if (!fields->foundPlayerOrAudioEffectsLayer) fields->foundPlayerOrAudioEffectsLayer = (node == this->m_player1 || node == this->m_player2 || typeinfo_cast<AudioEffectsLayer*>(node));
+				if (!fields->foundPlayerOrAudioEffectsLayer) continue;
+				const auto particle = typeinfo_cast<CCParticleSystemQuad*>(node);
 				if (!particle) continue;
 				if (particle->getZOrder() == 39 || particle->getZOrder() == 61) {
 					/*
@@ -155,19 +155,18 @@ class $modify(MyPlayLayer, PlayLayer) {
 				}
 			}
 		}
-		if (!m_fields->performHideTwoPlayerGuide) return;
+		if (!fields->performHideTwoPlayerGuide) return;
 		if (Utils::getBool("hideTwoPlayer")) {
 			/*
 			"omg ery why don't you hook PlayLayer::showTwoPlayerGuide()!! bad geode mod dev!"
 			simple: refusing to call the original will cause more issues beyond intended
 			also no bindings addresses!!
 			*/
-			for (auto node : CCArrayExt<CCNode*>(getChildren())) {
-				if (!m_fields->foundHitboxNodeTwoPlayerGuide) m_fields->foundHitboxNodeTwoPlayerGuide = node->getID() == "hitbox-node";
-				if (!m_fields->foundHitboxNodeTwoPlayerGuide) continue;
-				if (auto text = typeinfo_cast<CCLabelBMFont*>(node)) {
-					std::string labelString = std::string(text->getString());
-					if (labelString != "Player 1" && labelString != "Player 2") continue;
+			for (CCNode* node : CCArrayExt<CCNode*>(getChildren())) {
+				if (!fields->foundHitboxNodeTwoPlayerGuide) fields->foundHitboxNodeTwoPlayerGuide = node->getID() == "hitbox-node";
+				if (!fields->foundHitboxNodeTwoPlayerGuide) continue;
+				if (const auto text = typeinfo_cast<CCLabelBMFont*>(node)) {
+					if (const auto &labelString = static_cast<const std::string&>(text->getString()); labelString != "Player 1" && labelString != "Player 2") continue;
 					text->setVisible(false);
 					text->setScale(0.0f);
 					text->setContentHeight(0.0f);
@@ -178,10 +177,10 @@ class $modify(MyPlayLayer, PlayLayer) {
 					text->setString("");
 					text->setPositionX(PREFERRED_HOOK_PRIO);
 					text->setPositionY(PREFERRED_HOOK_PRIO);
-				} else if (auto sprite = typeinfo_cast<CCSprite*>(node)) {
+				} else if (const auto sprite = typeinfo_cast<CCSprite*>(node)) {
 					if (typeinfo_cast<CCLabelBMFont*>(sprite)) continue;
 					if (cocos::isSpriteFrameName(sprite, "floorLine_001.png")) sprite->setVisible(false);
-					m_fields->foundHitboxNodeTwoPlayerGuide = false;
+					fields->foundHitboxNodeTwoPlayerGuide = false;
 					break;
 				}
 			}
@@ -192,12 +191,14 @@ class $modify(MyPlayLayer, PlayLayer) {
 		return PlayLayer::toggleGlitter(p0);
 	}
 	void onQuit() {
-		m_fields->performHideTwoPlayerGuide = false;
-		m_fields->foundPlayerOrAudioEffectsLayer = false;
-		m_fields->foundHitboxNodeTwoPlayerGuide = false;
-		m_fields->foundHitboxNode = false;
-		m_fields->manager->endPortalX = 0.0f;
-		m_fields->manager->endPortalY = 0.0f;
+		Manager* manager = Manager::getSharedInstance();
+		const auto fields = m_fields.self();
+		fields->performHideTwoPlayerGuide = false;
+		fields->foundPlayerOrAudioEffectsLayer = false;
+		fields->foundHitboxNodeTwoPlayerGuide = false;
+		fields->foundHitboxNode = false;
+		manager->endPortalX = 0.0f;
+		manager->endPortalY = 0.0f;
 		PlayLayer::onQuit();
 	}
 	void startGame() {